+++ private/pkg/git/git.go
@@ -15,8 +15,14 @@
 package git
 
 import (
+	"bufio"
+	"bytes"
 	"context"
+	"errors"
+	"fmt"
+	"os/exec"
 	"regexp"
+	"strings"
 
 	"github.com/bufbuild/buf/private/pkg/app"
 	"github.com/bufbuild/buf/private/pkg/command"
@@ -24,6 +30,23 @@ import (
 	"github.com/bufbuild/buf/private/pkg/storage/storageos"
 	"github.com/bufbuild/buf/private/pkg/tracing"
 	"go.uber.org/zap"
+)
+
+const (
+	gitCommand      = "git"
+	gitOriginRemote = "origin"
+	tagsPrefix      = "refs/tags/"
+	headsPrefix     = "refs/heads/"
+)
+
+var (
+	// ErrRemoteNotFound is returned from GetRemote when the specified remote name is not
+	// found in the current git checkout.
+	ErrRemoteNotFound = errors.New("git remote not found")
+
+	// ErrInvalidGitCheckout is returned from CheckDirectoryIsValidGitCheckout when the
+	// specified directory is not a valid git checkout.
+	ErrInvalidGitCheckout = errors.New("invalid git checkout")
 )
 
 // Name is a name identifiable by git.
@@ -134,4 +157,200 @@ type ListFilesAndUnstagedFilesOptions struct {
 	// These must be unnormalized in the manner of the local OS that the Lister
 	// is being applied to.
 	IgnorePathRegexps []*regexp.Regexp
+}
+
+// Remote represents a Git remote and provides associated metadata.
+type Remote interface {
+	// Name of the remote (e.g. "origin")
+	Name() string
+	// HEADBranch is the name of the HEAD branch of the remote.
+	HEADBranch() string
+	// Hostname is the host name parsed from the remote URL. If the remote is an unknown
+	// kind, then this may be an empty string.
+	Hostname() string
+	// RepositoryPath is the path to the repository based on the remote URL. If the remote
+	// is an unknown kind, then this may be an empty string.
+	RepositoryPath() string
+	// SourceControlURL makes the best effort to construct a user-facing source control url
+	// given a commit sha string based on the remote source, and available hostname and
+	// repository path information.
+	//
+	// If the remote hostname contains bitbucket (e.g. bitbucket.mycompany.com or bitbucket.org),
+	// we construct the source control URL as:
+	//
+	//   https://<hostname>/<repository-path>/commits/<git-commit-sha>
+	//
+	// If the remote hostname contains github (e.g. github.mycompany.com or github.com), we
+	// construct the source control URL as:
+	//   https://<hostname>/repository-path>/commit/git-commit-sha>
+	//
+	// If the remote hostname contains gitlab (e.g. gitlab.mycompany.com or gitlab.com), we
+	// construct the source control URL as:
+	//   https://<hostname>/repository-path>/commit/git-commit-sha>
+	//
+	// If the remote is unknown and/or no hostname/repository path information is available,
+	// this will return an empty string.
+	//
+	// This does not do any validation against the gitCommitSha provided.
+	SourceControlURL(gitCommitSha string) string
+
+	isRemote()
+}
+
+// GetRemote gets the Git remote based on the given remote name.
+// In order to query the remote information, we need to pass in the env with appropriate
+// permissions.
+func GetRemote(
+	ctx context.Context,
+	runner command.Runner,
+	envContainer app.EnvContainer,
+	dir string,
+	name string,
+) (Remote, error) {
+	return getRemote(ctx, runner, envContainer, dir, name)
+}
+
+// CheckDirectoryIsValidGitCheckout runs a simple git rev-parse. In the case where the
+// directory is not a valid git checkout (e.g. the directory is not a git repository or
+// the directory does not exist), this will return a 128. We handle that and return an
+// ErrInvalidGitCheckout to the user.
+func CheckDirectoryIsValidGitCheckout(
+	ctx context.Context,
+	runner command.Runner,
+	envContainer app.EnvContainer,
+	dir string,
+) error {
+	stdout := bytes.NewBuffer(nil)
+	stderr := bytes.NewBuffer(nil)
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("rev-parse"),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+		command.RunWithEnv(app.EnvironMap(envContainer)),
+	); err != nil {
+		var exitErr *exec.ExitError
+		if errors.As(err, &exitErr) {
+			if exitErr.ProcessState.ExitCode() == 128 {
+				return fmt.Errorf("dir %s: %w", dir, ErrInvalidGitCheckout)
+			}
+		}
+		return err
+	}
+	return nil
+}
+
+// CheckForUncommittedGitChanges checks if there are any uncommitted and/or unchecked
+// changes from git based on the given directory.
+func CheckForUncommittedGitChanges(
+	ctx context.Context,
+	runner command.Runner,
+	dir string,
+) ([]string, error) {
+	stdout := bytes.NewBuffer(nil)
+	stderr := bytes.NewBuffer(nil)
+	var modifiedFiles []string
+	// Unstaged changes
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("diff", "--name-only"),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+	); err != nil {
+		return nil, err
+	}
+	modifiedFiles = append(modifiedFiles, getAllTrimmedLinesFromBuffer(stdout)...)
+
+	stdout = bytes.NewBuffer(nil)
+	stderr = bytes.NewBuffer(nil)
+	// Staged changes
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("diff", "--name-only", "--cached"),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+	); err != nil {
+		return nil, err
+	}
+
+	modifiedFiles = append(modifiedFiles, getAllTrimmedLinesFromBuffer(stdout)...)
+	return modifiedFiles, nil
+}
+
+// GetCurrentHEADGitCommit returns the current HEAD commit based on the given directory.
+func GetCurrentHEADGitCommit(
+	ctx context.Context,
+	runner command.Runner,
+	dir string,
+) (string, error) {
+	stdout := bytes.NewBuffer(nil)
+	stderr := bytes.NewBuffer(nil)
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("rev-parse", "HEAD"),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+	); err != nil {
+		return "", err
+	}
+	return strings.TrimSpace(stdout.String()), nil
+}
+
+// GetRefsForGitCommitAndRemote returns all refs pointing to a given commit based on the
+// given remote for the given directory. Querying the remote for refs information requires
+// passing the environment for permissions.
+func GetRefsForGitCommitAndRemote(
+	ctx context.Context,
+	runner command.Runner,
+	envContainer app.EnvContainer,
+	dir string,
+	remote string,
+	gitCommitSha string,
+) ([]string, error) {
+	stdout := bytes.NewBuffer(nil)
+	stderr := bytes.NewBuffer(nil)
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("ls-remote", "--heads", "--tags", "--refs", remote),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+		command.RunWithEnv(app.EnvironMap(envContainer)),
+	); err != nil {
+		return nil, err
+	}
+	scanner := bufio.NewScanner(stdout)
+	var refs []string
+	for scanner.Scan() {
+		line := strings.TrimSpace(scanner.Text())
+		if ref, found := strings.CutPrefix(line, gitCommitSha); found {
+			ref = strings.TrimSpace(ref)
+			if tag, isTag := strings.CutPrefix(ref, tagsPrefix); isTag {
+				refs = append(refs, tag)
+				continue
+			}
+			if branch, isBranchHead := strings.CutPrefix(ref, headsPrefix); isBranchHead {
+				refs = append(refs, branch)
+			}
+		}
+	}
+	return refs, nil
+}
+
+func getAllTrimmedLinesFromBuffer(buffer *bytes.Buffer) []string {
+	scanner := bufio.NewScanner(buffer)
+	var lines []string
+	for scanner.Scan() {
+		lines = append(lines, strings.TrimSpace(scanner.Text()))
+	}
+	return lines
 }